More search improvements, got some filter interface prototyped.
This right now works quite fine if you do some query like:

	`search.search('author:admin', filter=PostDocumentType().get_filter(User.objects.get(username='admin')))

Need to polish this a bit more and need to find a way to join up more filters from different
DocumentTypes.  But hey, this stuff is right now so damn crazy fast and we are doing some
advanced stuff already.