Added very first graphical interface for the new search, still very experimental
Splitted `search` to `parse_query` to actually give the user interface more
freedom to form a proper query.

Added autodiscover for the initial import so that the search indices
are build when everything is loaded (speeds up request cycle).

Added special boost query support using dis_max_query, this replaces the
current "This wiki page is exactly what you are looking for" implementation
but does this for all indices.  So that whenever we have an exact match
it is boosted to the top.