Make it possible for admins to edit any public user profile.
This change is very redundant, basically only c&p, but so is the whole
usercp / admincp. I guess it would make sense to unify usercp and admin
editing functions, but this would go to far for this one pull request
and takes too much time currently.
But because making it possible for administrators to change any user
profile because of spam / illegal content / etc. is essential, I
implemented this "quick and dirty" instead of redesigning the usercp /
admincp to not delay the whole branch and making it difficult to merge.