Implemented inyoka.core and added some new autodiscover signal for special modules.
This imports modules like '.search' for every application and in the end allows
to implement pluggable APIs where a application can register addons to some
core functionality.

With this we are able to finally split application logic from core functionality
and cleanup some fancy circular imports.

This truely adds some overhead for the initial startup time
of Inyoka but as Python caches imports the overhead is bearable.