Removed old, unused and unmaintained generate_markup functionality.
This was used to initially migrate the old ubuntuusers wiki syntax over
to Inyoka.

As for many nodes the markup representation is unmaintained and not
complete and unused as well I remove it as I do not see any real
benefit in it.

Some might argue that with this we can add stuff like PDF-import and more
sure, but in the real world I would prefer wrapping this with [[HTML()]]
and use one of the might _to_html converters out there as many major
wiki engines do.

This also finally allows for multiple macro names to complete i18n.