Started to refactor macros to allow pluggable registration and multiple aliasses.
Implemented a similar API to our search system where every application can plug-in
it's own implementation. This allows for cleaner application structure and minimizes
circular imports.

With this commit we also finally support multiple aliases for macros. This means
in fact that we do support the english names of all macros and this finally
allows us to go much further on i18n as we could ever before. Still much
work to do here obviously.