Removed UserManager.get()
You have to use .get(username__iexact="<the name>") instead of .get("<the name>"). This is a BACKWARDS INCOMPATIBLE change.

The changes in
* inyoka/utils/search.py
* inyoka/utils/forms.py
will change the behavoir a little bit.

The username wont be normalized in UserManager.get anymore you have to replace underscores with tabs yourself now (better: use normalize_username().