Change the max_length to 255 for metadata values
For an InnoDB 255 is the max key length, if uft-8 is used. Because it
assumes that every char needs the maximum of 3 bytes. An key can not be
longer than 767 bytes, so you can not set more than 255 chars.

Since this limit is only valid for the indexed keys another fix would be
not to index this field.