Fix: prioritize username over email.
We had a user today who registered with his e-mail address as username but
then changed his mind, closed his account and registered again using a
different username but the same e-mail address.

Trying to edit the old user account will redirect to the new user account,
because `get_user()` will prioritize e-mail addresses if there is an `@` in
username.

We already have a method on the UserManager that gets a user object based on
the username or email address. So let's drop the redundant code altogether.
This method will also try to use the username first and then fall back to
e-mail addresses, solving the issue above.