Replace login requierment check with Django logic
There we two functions which checked if the user was logged in.
Because Django also provides a simular function, this is used
instead of an own implementation. Unfortunaly the url resolver
service is also an inyoka implementation which is not compitable
with the Django resolve. Because of that we need to provide the
login url for each check.