Fix wrong variable name in create article form
Clear oversight on my part. When creating new articles, if the user does
not have the right to create the article, it's supposed to be created in a
construction area instead. Since I used the wrong variable name, this never
returned the name of the construction area and it never worked as intended.