Add configurable help text to wiki/create view.
For internationalization we should always link to the main create view.

Up until recently, we had a macro to embed in a wiki article – this article
was linked to based on a setting `wiki_newpage_infopage`.
This drops the old setting in favor of a configurable help text, to be
displayed on the create view.

(See also changes to templates)