Allow redirects to use anchors.
As @chris34 pointed out to me on IRC, redirects can have an anchor.
`Page.objects.get_by_name()` doesn't expect the name to contain an anchor and
this could potentially cause issues.

Anchors are now handled by `get_safe_redirect_target`. That means it will
ignore anchors on the first redirect and use the anchor of the final redirect
(if present). It returns a tuple with the target slug and the anchor name.

I added a test for anchors and made the two changes suggested by @ostcar.