Updated all urls.py files to the new format.
The documentation of django recommands to use the url function to define
urls.

Besides django changed the format of the urlpattern object that is expected
in the urls.py module. The support for the old format will be dropped in
django 1.10.

Now, the second argument of the url function has to be a callable.

See https://docs.djangoproject.com/en/1.9/topics/http/urls/

I fixed the most of it with this regular expression:
^(\s*)(?:url|)(\(r'.*?',\s*)'(.*?)'