Rework on forum welcome message and show supporter of group.
Each forum has supporters. Before this commit this where all users that had
permissions to moderate the forum. With this commit each forum can have a
supporter group and any user of this group is listed as supporter. This
is an important step to remove user permissions.

This commit is also a rework of the forum welcome message, because the new
implementation is easier to maintain.

* The ForumEditForm to create or update a forum is now a ModelForm.
* The welcome text and welcome title are now fields of the forum model instead of an own model
* The information which user has accepted a forum is now a many2many table instead of a TextField
* The view to edit a forum is now implemented with django class based views
* The PermissionMixin for class based views from django 1.9 was added for the time we use django 1.8

The migration filed added by this commit do only alter the table but do not
migrate the data. Therefore the welcome text and welcome title have to be
updated manualy and any user has to reaccept the message.