Create the possibility to run tests on different database servers which do not run on localhost, at the same time.
To run tests there have to be different database servers. The tests try to connect
to this servers on localhost as default. To specify a different host, a file
custom.py has to be created.

Before this commit the custom file had to be used as settings, so it could not
be configured for different databases at the same time. With this commit, the
tests are started my choosing not the custom-settings, but the "normal" setting
like tests.settings.mysql or tests.settings.postgresql.

See the created README for more information