Fix the forum post_counter.
There was an inconsistency. When the post_counter of a forum was recalculated
then only the posts of the forum where counted and not the posts of the
child forums. But if a new post was created in a child forum, then the post_counter
of the parent forum was increased.

Now the posts of child forums are allways counted.

Also added the property forum.descendants that returns not only the children
of a forum but all descendants.

Also fixed a small bug in the QueryCounter class. If the QueryCounter was recalculated
with get_or_set then the value was calculated only once but redis.set was called
twice.