Rewrote the WelcomeMessageView as class based view.
Also made the redirect from an unaccepted forum to the welcome message
page is now a redirect (http 302 response) instead of a internal
redirect. On the pro side, the user gets the right url. On the con side,
after accepting the forum the user gets always redirected to the forum
and not to the side he was comming from (for example a topic).

The reason for the con side is, that it is a security problem, if the user can
set the url where the welcome message page redirects to.