Fix on_toggle_sidebar() Service.
If, for any possible or inpossible reason, a Users settings Dictionary (User.settings) doesn't
have an field, we should Default pop() Operations to avoid KeyErrors.

This happened on http://sentry.lan-cgn:9000/ubuntuusers/production/group/169/. Using a check with
has_key() is possible, but wastes time for things we dont need.