Small refactor of User model to use constants instead of literals.
`user.status` determines if the user is banned, or inactive, etc. We changed or
checked it's value throughout the code base using literals. `user.status = 2`.

In this change I switched it to constant values that are easy to understand.
We also had some properties on the model that weren't used, instead we manually
checked against the literal values.

I also rewrote the properties to avoid lambda functions, better for readability.