Create System Groups (registered, anonymous, ikhaya) on demand.
The three system groups (registered, anonymous, ikhayateam) are now created
on demand. This ensures that we have these always available.

This commit changes also from and ID based specification of system groups
to and name based one. This ensures that not any funny named group from
make_testdate gets any special meanings in the future.

In existing setups it is require to change settings.py accordingly.