Fix: deleted wiki discussions cause server errors
Deleting forum topics would not reset the cache of wiki pages, if the topic was
marked as a discussion for said wiki page. Until the cache is reset, such pages
would trigger a `DoesNotExist` exception.

This commit ensures that cache keys for wiki pages are deleted when topics are
deleted.