Ensure that all active users are always in the registered group.
This means for make_testdata if status==1 we assign the registered
group. On production systems we use an migration via RunSQL() for it.

A RAW SQL is needed, because iterating over ordinary python objects is
really slow (several minutes, on postgresql without fsync). The RAW query
only needs a few seconds to complete.