forum: splittopic properly handle url based slugs
Previously the check for url based slugs was nasty and just took the last part of the URL. This does not work for topic URLs with pages, the detected slug would be the page number!
Now we use proper urlparsing using the lib and extract the slug using the absolute position.