Remove portal_sessioninfo for performance reasons.
According to a analysis it did today portal_sessioninfo is the most
written table in our database. >50% of all write request on production
go to this table.

The only known reason for this table is the user record and information
about logged in users in a portal view. I moved this tasks and the
sessioninfos to redis, without changing the behavior.