Test the planets feed functions with BBD tests
Covered by this tests is the view permissions for the post in the feed
and the functionality to hide and restore a blog post in the feed.

A minor bug was also fixed: If a blog post was restored the hidden_by user
field remained on its value. Now it gets reset to None.

The tasks.py was excluded from the coverage because reading in a feed
should be tested by a unit test with mocks and not by a selenium test.