Post revisions: Order by date via DB
Instead of reversing the objects from the DB with Python, we let the DB
do the ordering. That should be still faster, even if there is no index for
store_date. IMO the overhead for an index is not needed, as only
moderators will notice that at all. (The current text of a post is
always stored directly inside `Post`)

After the postgresql-migration the ordering could also break without this
change.