Let the UI tests run again with Django 1.11
This commit updates the dependencies and adjusts the UI-Test so that they
run in general.

Django 1.11 changed the test setting an the behave-djnago project didn't
allow any setting of the host name anymore. Therefore the dependency needed
to be removed and the setup was switched back to the plain behave set up.

Also the support for PhantomJS was deprecated with the newest Selenium
version therefor the tests browser was switched to chromium.