Use Emoji instead of smileymap
Instead of a configurable smileymap that was editable via a wikipage,
there is now a hardcoded dict SMILIES in the settings. Latter includes
the inyoka ASCII-markup (like `:)`) and the corresponding emoij replacement
(like 😀). If there is no equivalent in unicode, a image sprite via a <span>
with a special css-class is used in the frontend.

The reason for omitting the smileymap is that it is now possible to use
UTF8-emojis in every textfield after the migrations to postgresql.
Nevertheless, the hardcodede dict preserves legacy behaviour.