@transaction.atomic for mv_* wiki actions.
Currently (with PostgreSQL) we have some issues of mv/rename in our
wiki, this can lead to database corruption if the request/worker
gets killed (for timeour reasons) during operation.

With this change, the whole task is one DB transaction, so we have
a broken cache in the worst case, but a still valid database!