Wiki: Prepend newpage_root for wikisubpages
For normal users that are only allowed to create pages in the
newpage_root (on uu.de: 'Baustelle') the newpage_root was not
automatically prefixed, if a subpage (like `Howto/foobar`) should be created.
It was possible for them to create `foobar` and also
`Baustelle/Howto/foobar`, if latter was manually inserted.
See https://forum.ubuntuusers.de/topic/benutzer-kann-kein-neues-howto-erstellen/#post-9071388

`join_pagename` uses the second argument, if the second argument contains `/` and
does not start with a `./`. (IMO the semantic should be same as
`os.path.join` completly, but it has to be searched, where this change breaks code)

Tests are provided to harden against a regression.