Prevent deletion of teamicon, if users edit their profile
Problem: The team icon was always lost, if a team member edited
their own profile.

This was caused by an inheritance-problem in combination with
unrendered formfields in the frontend.

In more detail:
 * via `EditUserProfileForm` privileged users can edit other user
   profiles.
 * via `UserCPProfileForm` users can edit their own profile

`EditUserProfileForm` is a subclass of `UserCPProfileForm`.
Before the commit, `UserCPProfileForm` included the modelfield `icon` (the field
where the team icon of a user is saved). However, the formfield was not
rendered in the frontend. Thus, if `UserCPProfileForm` was submitted, `icon`
got the value `None` and the user lost its team icon.

As a solution, the logic for `icon` was moved from `UserCPProfileForm`
to `EditUserProfileForm`.