Own template for ImageCaptchaWidget
With Django 1.11 every widget got an own template, so it.
To ease configuraion, the default of django template engine is used for
this template to render. Jinja can be used for this template, too, if
Django's native template engines is used, see #983

A testcase was added to see, whether the captacha-image is rendered.

DateTimeWidget.render without the renderer-parameter is also deprecated,
so it was adapted.