Remove jsi18n-url
The javascript_catalog function-view is depcrated since Django
1.10 and removed with 2.0
https://docs.djangoproject.com/en/3.0/releases/2.0/#features-removed-in-2-0

Furthermore, it is not used in the frontend
(`ag 'pluralidx|gettext|ngettext|gettext_noop|pgettext|npgettext|interpolate|get_format'`
did find nothing in the theme)
and also fetched on *every request* (no caching…).

So i decided to drop this completly. If we need it, it can be readded
with the new class following the django docs.