Introduce ForumMangager.update_last_post to prevent last_post=None in categories
The new method is used at every place, where `last_post` of a Forum
has to be updated (move topic, split a topic etc.). So this logic is now
always the same.

The method affects multiple forum instances, so the manager
should be the correct place.

It uses Topic.last_post where appropriate. This makes the aggregate-query
in a production like setting, where many posts exists, faster.
If speed is no matter, in theory it should be also possible to just use an
aggreagte on Posts.