TestPostMove: Check last_post of parent and category-forums, too
Sentry reported a problem with a Typererror in the read status.

TypeError: unorderable types: int() >= NoneType()
  File "django/core/handlers/exception.py", line 41, in inner
    response = get_response(request)
  File "django/core/handlers/base.py", line 187, in _get_response
    response = self.process_exception_by_middleware(e, request)
  File "django/core/handlers/base.py", line 185, in _get_response
    response = wrapped_callback(request, *callback_args, **callback_kwargs)
  File "inyoka/forum/views.py", line 1558, in markread
    mark_all_forums_read(user)
  File "inyoka/forum/models.py", line 1485, in mark_all_forums_read
    user._readstatus.mark(forum, user)
  File "inyoka/forum/models.py", line 1429, in mark
    if self(item):
  File "inyoka/forum/models.py", line 1418, in __call__
    if row[0] and row[0] >= post_id:

The top categories have (sometimes) a `last_post` of `None` which causes
the traceback. The forum.id is f.e. 76, 58, 57, 61, 63 on the production
database.

F.e. in `Topic.move` the `last_post` is updated via something like

    forum.last_post_id = Post.objects.filter(topic__forum=forum).aggregate(count=Max('id'))['count']

for all parent forums. The aggregate can return `None`, if there is
no topic in the forum. This applies (at least) for all category forums.
A category forum is one with no parent forums and no topics, is has just subforums.

This is also reproducable via the added test. With this commit the test fails with

Traceback (most recent call last):
  File "inyoka/tests/apps/forum/test_models.py", line 260, in test_topic_move__topic_in_forum_left
    self.assertEqual(self.category.last_post, second_topic_posts[-1])
AssertionError: None != <Post id=12 author=3>