Upgrade requirements and rebuild the files with python 3.9
Unpin more requirements as far as they support the currently used
django-version.

As python 3.9 is used in the docker setup, the requirements are rebuild
with that python-version. It has only changed for the unpinned requirements,
so it can be also used for python 3.8.
Thus, production.txt and linux-py3.8-production.txt point to
linux-py3.9-production.txt now (likewise for the development files)