Add caching to every feed url
The views will be pickled and cached in redis-DB 0
('default' in the `inyoka.default_settings`).
The associated keys will start with `views.decorators.cache`
like in the example below.

```
/data # redis-cli

[…]

127.0.0.1:6379[2]> SELECT 1
OK
127.0.0.1:6379[1]> KEYS *
1) ":1:views.decorators.cache.cache_header..3e83b009139583fc354be837e64b4e0d.de-DE"
2) ":1:storage/global_message_rendered"
3) ":1:storage/session_record"
4) ":1:views.decorators.cache.cache_page..GET.3e83b009139583fc354be837e64b4e0d.d41d8cd98f00b204e9800998ecf8427e.de-DE"
5) ":1:forum/forums/einrichten-verwalten"
6) ":1:wiki/page/wiki/vorlagen/tasten"
7) ":1:forum/slugs"
8) ":1:/acl/1"
```

As the view's content is cached for 5 minutes, some hidden items
(for example a hidden forum post) could still be public in the feeds
for 5 minutes.
This should be acceptable, as most of the time a post with SPAM
will be also first public and after some time hidden by moderators.