Strip control characters in Inyoka markup lexer
After the migration to django feeds, some RSS feeds raised an
`UnserializableContentError`. An example traceback from sentry:

```
UnserializableContentError: Control characters are not supported in XML 1.0
  File "django/core/handlers/exception.py", line 47, in inner
    response = get_response(request)
  File "django/core/handlers/base.py", line 181, in _get_response
    response = wrapped_callback(request, *callback_args, **callback_kwargs)
  File "django/utils/decorators.py", line 130, in _wrapped_view
    response = view_func(request, *args, **kwargs)
  File "django/contrib/syndication/views.py", line 47, in __call__
    feedgen.write(response, 'utf-8')
  File "django/utils/feedgenerator.py", line 304, in write
    self.write_items(handler)
  File "django/utils/feedgenerator.py", line 338, in write_items
    self.add_item_elements(handler, item)
  File "django/utils/feedgenerator.py", line 370, in add_item_elements
    handler.addQuickElement("summary", item['description'], {"type": "html"})
  File "django/utils/xmlutils.py", line 20, in addQuickElement
    self.characters(contents)
  File "django/utils/xmlutils.py", line 27, in characters
    raise UnserializableContentError("Control characters are not supported in XML 1.0")
```

So, some user content seems to contain control characters.
An affected post in the forum is https://forum.ubuntuusers.de/post/8997239/

Thus, with this commit control characters are ignored in Inyoka's markup,
as they are also not allowed in HTML.
(see https://www.w3.org/International/questions/qa-controls
 that is also referenced in django's code before the error is raised)

W3C's validator also displays an error for the control characters, if
the example URL is submitted.