Add paths for template and statics of custom theme
The statics and templates of the custom theme should take precedence.
However, if files are not in a custom theme, the builtin Inyoka files
should be used as a 'fallback'.

Thus we have to adapt `STATICFILES_DIRS` and `TEMPLATES`.
This is done for the `example_development_settings.py`.

`findstatic` can help to determine, if it works for the static files. It should look like:

```
python manage.py findstatic style/css_grid.less
Found 'style/css_grid.less' here:
  /inyoka/theme-ubuntuusers/inyoka_theme_ubuntuusers/static/style/css_grid.less
  /inyoka/inyoka/static/style/css_grid.less
```