Restrict user defineable font faces
Before this change, every possible font family could be used by a user.
Only occurences of `[font=Arial]`, [font=serif]` and `[font=Courier]`
where found in the production database of ubuntuusers.de.

As this is only kept for backwards compatibility, we now
restrict the values that can be used.

This should increase

 - accessibility (no random font making text hard to read)
 - and security (XSS harder).