Login: Increase `max_length` of username field for mail addresses
The django form sets `max_length` of the form field to `max_length` of the
`username` field in the `User` *model*.

However, in Inyoka we also allow to enter email adresses as usernames.
These have a higher `max_length`.

Now the maximum of the email and username model-fields is used
(which is at the moment the value of the email-field).