Event: Slug is changed for unpublished events
Now, if the name of a hidden event is changed, the slug will be also
changed.

If the event is published, the slug will not change.
Mainly, to keep URLs stable.
(To intentionally change the slug of a published event, f.e. if it
contains sensitive information, the event can be temporarily hidden to
change the slug.)

Fix https://github.com/inyokaproject/inyoka/issues/520