Forum, Topic: Make slug unique (which creates an index)
Sentry reported a slow DB query for `Topic.slug`.
At the moment a full table scan is performed.
To fasten things up, the database field is marked as unique.
As a side effect, this will create an index to speed up lookups.

With this commit, on sqlite the explain changes

from
```
SCAN forum_topic
```

to
```
SEARCH forum_topic USING INDEX sqlite_autoindex_forum_topic_1 (slug=?)
```

Even if it was not enforced in the database, the slug is unique across
all topics on forum.ubuntuusers.de.