Remove ability to toggle admin links
Mostly caused trouble in the past that people asked why they dont
see admin links.

The class `admin_link_hover` was just removed in the JS before,
so remove it in the HTML, too.

The class `admin_link` was replaced with `admin_link_js` in the JS,
as the JS is removed, we can drop all its styles.

`#admin_layer_button` is removed, so we can drop all its styles.