Simplify redirect logic for ikhaya article subscribe/unsubscribe
The link is only used on
`inyoka/ikhaya/jinja2/ikhaya/detail.html` and
`inyoka/ikhaya/jinja2/ikhaya/index.html`.

Redirecting to the article is the default. If a user subscribes to
an article on the ikhaya index page, it should also redirect there.
This is now signalled by a simple `?next=index`.

So we dont possily redirect to an unwanted external location.