Wiki: Replace cancel button with link
The link is styled similar to a button.

Before, it was necessary to add a change comment, even if editing
should be canceled.

If there is something changed in the form, but not submitted,
javascript on the clientside will still show a warning.

Fix https://github.com/inyokaproject/inyoka/issues/1161