Replace custom `cleanup_html` with nh3
This also changes the approach to an allow list of HTML-elements.
Additionally, lxml build issues should not be a problem anymore.
(For `<img> the attributes `srcset` & `sizes` are not allowed to simplfy the implementation)

In inyoka.utils.html unused methods and logic were removed.
For example `<img />` is not used in HTML5 anymore. `<img>` is enough.
The end tag was generated in `_build_html_tag`, but discarded in `build_html_tag`.